+++ doc/crypto/api.db/psa/crypto.h
@@ -67,6 +67,10 @@ typedef struct psa_custom_key_parameters_t {
 #define PSA_ALG_ASCON_XOF128 ((psa_algorithm_t)0x0D000300)
 #define PSA_ALG_AT_LEAST_THIS_LENGTH_MAC(mac_alg, min_mac_length) \
     /* specification-defined value */
+#define PSA_ALG_BLAKE2B_HASH512 ((psa_algorithm_t)0x0200001B)
+#define PSA_ALG_BLAKE2B_MAC ((psa_algorithm_t)0x03800400)
+#define PSA_ALG_BLAKE2S_HASH256 ((psa_algorithm_t)0x0200001A)
+#define PSA_ALG_BLAKE2S_MAC ((psa_algorithm_t)0x03800300)
 #define PSA_ALG_CBC_MAC ((psa_algorithm_t)0x03c00100)
 #define PSA_ALG_CBC_NO_PADDING ((psa_algorithm_t)0x04404000)
 #define PSA_ALG_CBC_PKCS7 ((psa_algorithm_t)0x04404100)
@@ -337,6 +341,7 @@ typedef struct psa_custom_key_parameters_t {
 #define PSA_KEY_TYPE_ARC4 ((psa_key_type_t)0x2002)
 #define PSA_KEY_TYPE_ARIA ((psa_key_type_t)0x2406)
 #define PSA_KEY_TYPE_ASCON ((psa_key_type_t)0x2008)
+#define PSA_KEY_TYPE_BLAKE2 ((psa_key_type_t)0x200B)
 #define PSA_KEY_TYPE_CAMELLIA ((psa_key_type_t)0x2403)
 #define PSA_KEY_TYPE_CHACHA20 ((psa_key_type_t)0x2004)
 #define PSA_KEY_TYPE_DERIVE ((psa_key_type_t)0x1200)